+++ src/minecraft/micdoodle8/mods/galacticraft/moon/dimension/GCMoonWorldProvider.java
@@ -14,7 +14,7 @@ import cpw.mods.fml.relauncher.SideOnly;
 import cpw.mods.fml.relauncher.Side;
 
 /**
- * Copyright 2012, micdoodle8
+ * Copyright 2012-2013, micdoodle8
  * 
  *  All rights reserved.
  *
@@ -63,7 +63,7 @@ public class GCMoonWorldProvider extends WorldProvider implements IGalacticraftW
 	@Override
     public Vec3 getSkyColor(Entity cameraEntity, float partialTicks)
     {
-        return this.worldObj.getWorldVec3Pool().getVecFromPool((double)0, (double)0, (double)0);
+        return this.worldObj.getWorldVec3Pool().getVecFromPool(0, 0, 0);
     }
 	
 	@Override
